Split Parser Tests
